roll back to old facebook comments
