+++ package.json
@@ -57,7 +57,6 @@
     "rc-slider": "^11.1.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-facebook": "^9.0.12",
     "react-helmet": "^6.1.0",
     "react-router-dom": "5",
     "react-transition-group": "^4.4.5"
